PAE-1325: fold get-status-tag-class into format-submission-status
