+++ src/server/reports/helpers/format-submission-status.test.js
@@ -1,8 +1,34 @@
 import { describe, it, expect } from 'vitest'
 
-import { getStatusLabel, getActionLabel } from './format-submission-status.js'
+import {
+  getActionLabel,
+  getStatusLabel,
+  getStatusTagClass
+} from './format-submission-status.js'
 
 const localise = (key) => key
+
+describe('#getStatusTagClass', () => {
+  it.each([
+    { status: 'due', expected: 'govuk-tag--orange' },
+    { status: 'in_progress', expected: 'govuk-tag--yellow' },
+    { status: 'ready_to_submit', expected: '' },
+    { status: 'submitted', expected: 'govuk-tag--green' }
+  ])(
+    'returns "$expected" modifier class for "$status" status',
+    ({ status, expected }) => {
+      expect(getStatusTagClass(status)).toBe(expected)
+    }
+  )
+
+  it('returns empty string for null status', () => {
+    expect(getStatusTagClass(null)).toBe('')
+  })
+
+  it('returns empty string for unrecognised status', () => {
+    expect(getStatusTagClass('unknown')).toBe('')
+  })
+})
 
 describe('#getStatusLabel', () => {
   it('returns localised label for "due" status', () => {
